requirements.txt
chromadb
sentence-transformers
groq
ollama
gradio
python-dotenv
